Cleaned up the passing parameters
