+++ currencyii/src/main/java/nl/tudelft/trustchain/currencyii/sharedWallet/SWTransferFundsAskTransactionData.kt
@@ -21,8 +21,9 @@ data class SWTransferFundsAskBlockTD(
 class SWTransferFundsAskTransactionData(data: JsonObject) : SWBlockTransactionData(
     data, CoinCommunity.TRANSFER_FUNDS_ASK_BLOCK
 )  {
-
+    // TODO: Update 0 and 1 to corresponding lists, but I can't find them yet
     var SW_VOTES:HashMap<Int, ArrayList<String>> = hashMapOf(0 to arrayListOf(), 1 to arrayListOf(), 2 to ArrayList(getData().SW_BITCOIN_PKS))
+
 
     fun getData(): SWTransferFundsAskBlockTD {
         return Gson().fromJson(getJsonString(), SWTransferFundsAskBlockTD::class.java)
@@ -30,7 +31,7 @@ class SWTransferFundsAskTransactionData(data: JsonObject) : SWBlockTransactionDa
 
     fun userVotes(userID: String, voteID: Int): HashMap<Int, ArrayList<String>> {
         SW_VOTES[2]!!.remove(userID)
-        SW_VOTES[voteID]!!.toMutableList().add(userID)
+        SW_VOTES[voteID]!!.add(userID)
         //TODO: Send this to others
         return SW_VOTES
     }
